Move grasp data struct to separate file
