+++ include/moveit_simple_grasps/grasp_data.h
@@ -1,0 +1,98 @@
+/*
+ * Software License Agreement (Modified BSD License)
+ *
+ *  Copyright (c) 2014, University of Colorado, Boulder, PAL Robotics, S.L.
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions
+ *  are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/or other materials provided
+ *     with the distribution.
+ *   * Neither the name of Univ of CO, Boulder, PAL Robotics, S.L.
+ *     nor the names of its contributors may be used to endorse or
+ *     promote products derived from this software without specific
+ *     prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ *  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ *  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ *  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *  POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/* Authors: Bence Magyar, Dave Coleman
+   Description: Data class used by the grasp generator.
+*/
+
+#ifndef MOVEIT_SIMPLE_GRASPS__GRASP_DATA_H_
+#define MOVEIT_SIMPLE_GRASPS__GRASP_DATA_H_
+
+// ROS
+#include <ros/ros.h>
+
+// TF
+#include <tf_conversions/tf_eigen.h>
+
+// Msgs
+#include <geometry_msgs/PoseArray.h>
+
+// MoveIt
+#include <moveit_msgs/Grasp.h>
+#include <moveit/macros/deprecation.h>
+
+// Eigen
+#include <Eigen/Core>
+#include <Eigen/Geometry>
+#include <eigen_conversions/eigen_msg.h>
+
+// Visualization
+#include <moveit_visual_tools/visual_tools.h>
+
+// C++
+#include <math.h>
+#define _USE_MATH_DEFINES
+
+namespace moveit_simple_grasps
+{
+
+struct RobotGraspData
+{
+  RobotGraspData() :
+    // Fill in default values where possible:
+    base_link_("/base_link"),
+    grasp_depth_(0.12),
+    angle_resolution_(16),
+    approach_retreat_desired_dist_(0.6),
+    approach_retreat_min_dist_(0.4),
+    object_size_(0.04)
+  {}
+  geometry_msgs::Pose grasp_pose_to_eef_pose_; // Convert generic grasp pose to this end effector's frame of reference
+  trajectory_msgs::JointTrajectory pre_grasp_posture_; // when the end effector is in "open" position
+  trajectory_msgs::JointTrajectory grasp_posture_; // when the end effector is in "close" position
+  std::string base_link_; // name of global frame with z pointing up
+  std::string ee_parent_link_; // the last link in the kinematic chain before the end effector, e.g. "/gripper_roll_link"
+  std::string ee_group_; // the end effector name
+  std::string ee_joint_; // the joint to actuate for gripping
+  double grasp_depth_; // distance from center point of object to end effector
+  int angle_resolution_; // generate grasps at PI/angle_resolution increments
+  double approach_retreat_desired_dist_; // how far back from the grasp position the pregrasp phase should be
+  double approach_retreat_min_dist_; // how far back from the grasp position the pregrasp phase should be at minimum
+  double object_size_; // for visualization
+};
+
+} // namespace
+
+#endif
